test: rename helper/css/common.test.tsx to helper/css/common.case.test.tsx
